Unit Overview upgrade - reorg
